mutakhir

+++ database/migrations/2022_07_08_060833_create_nilais_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
     {
         Schema::create('nilais', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('nilai');
+            $table->double('nilai');
             $table->timestamps();
             $table->softDeletes();
         });
